bt-849: * refactor logic
