+++ mobile/src/bundles/common-data/hooks/hooks.ts
@@ -1,2 +1,1 @@
 export { useCommonData } from './use-common-data/use-common-data';
-export { useLmsData } from './use-lms-data/use-lms-data';
