+++ mobile/src/bundles/talent/components/project-container/constants/constants.ts
@@ -1,6 +1,4 @@
-import { type TalentInfoProject } from '~/bundles/common/types/types';
-
-const PROJECT_MOCK: TalentInfoProject = {
+const PROJECT_MOCK = {
     details: {
         en: '6 weeks / 6 engineers, 2 QA / JS / Healthtech industry',
         ua: '',
